Convert homepage form to regular form
Does this mean no more React on what is basically a static website?

Also got gz paths working. I have a concept for an even more compressed string knowing what character types are possible. Basically turning the whole string into a list of enum values, but with an escape hatch for JSON since things like the meta and custom characters still need it. If that's not possible, then gz will always be a fallback.